edit security

+++ src/main/java/hexlet/code/controller/UserController.java
@@ -9,6 +9,8 @@ import hexlet.code.model.User;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.http.HttpStatus;
 
@@ -16,6 +18,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import hexlet.code.repository.UserRepository;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @RequestMapping("/api/users")
@@ -61,15 +64,32 @@ public class UserController {
         var user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
 
-        userMapper.map(userData, user);
-        userRepository.save(user);
+        // Проверка, что пользователь может редактировать только свои данные
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentPrincipalName = authentication.getName();
+        if (!user.getUsername().equals(currentPrincipalName)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You can only update your own data.");
+        } else {
+            userMapper.map(userData, user);
+            userRepository.save(user);
+        }
+
         return userMapper.map(user);
     }
-
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long id) {
-        userRepository.deleteById(id);
+        var user = userRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
+
+        // Проверка, что пользователь может удалять только себя
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentPrincipalName = authentication.getName();
+        if (!user.getUsername().equals(currentPrincipalName)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You can only delete your own data.");
+        } else {
+            userRepository.deleteById(id);
+        }
     }
 }
